backend/requirements.txt
databases==0.8.0
fastapi==0.104.1
passlib==1.7.4
pydantic==2.5.1
PyJWT==2.8.0
PyJWT==2.8.0
SQLAlchemy==1.4.50
uvicorn==0.15.0
python-dotenv==0.19.1
alembic
gunicorn
asyncpg
psycopg2-binary
pyTelegramBotAPI
requests
telebot
fpdf
